Update consumers for MjbDoubleSpan (float→double migration)
MjEngineTool: float* helpers → double*, MjbFloatSpan → MjbDoubleSpan,
remove MjbBackend parameter from LoadModel/LoadModelFromString,
add MjVector3AtEntry/MjQuaternionAtEntry raw double* overloads.

Components: add (float) casts where span indexer (now double) is stored
to float fields (MjHingeJoint, MjSlideJoint, MjActuator, MjBaseTendon,
MjSpatialTendon).

Tests: use SetQposAt/SetQvelAt instead of read-only span indexer writes.

+++ Runtime/Components/Joints/MjHingeJoint.cs
@@ -56,7 +56,7 @@ namespace Mujoco {
       RawConfiguration = data.GetQpos()[QposAddress];
       var qpos = (float)RawConfiguration;
       Configuration = (qpos % (2 * Mathf.PI)) * Mathf.Rad2Deg;
-      Velocity = data.GetQvel()[DofAddress];
+      Velocity = (float)data.GetQvel()[DofAddress];
     }
 
     protected override void OnParseMjcf(XmlElement mjcf) {
